好友-add-好友ID升级为int64

+++ Assets/Script/Game/Module/Character/CharacterModule.cs
@@ -36,7 +36,7 @@ namespace SGame
         /// <param name="id"></param>
         /// <param name="pos"></param>
         /// <returns></returns>
-        public CharacterSpawnResult Create(int id, Vector3 pos, bool hasAttribute = true, int playerID = 0)
+        public CharacterSpawnResult Create(int id, Vector3 pos, bool hasAttribute = true, long playerID = 0)
         {
             return CharacterSpawnSystem.CharacterSpawn.Create(id, pos, hasAttribute, playerID);
         }
